+++ src/app/index.js
@@ -25,13 +25,16 @@
     }
 
     // 启动逻辑
-    startLogic.$inject = ['userService', 'workbookService'];
-    function startLogic(userService, workbookService) {
+    startLogic.$inject = ['userService', 'workbookService', 'coreCF'];
+    function startLogic(userService, workbookService, config) {
+      // 用户初始化
       userService.initialize().then(function(user) {
-        var dime = user.record.dime;
+        var dime = user.record.dime; // 维度记录
 console.info(user);
+if (config.debug === true) { dime = config.dime; }
+        // 工作簿初始化
         workbookService.initialize(dime).then(function(workbook) {
-
+          workbookService.toggle(0, true);
         });
         // angular.forEach(dime, function(dim, index) {
         //   console.info(dim);