+++ src/app/workbook/workbook.controller.js
@@ -6,17 +6,24 @@
     .module('pf.workbook')
     .controller('WorkBookCtrl', workbookCtrl);
 
-  workbookCtrl.$inject = ['$scope', 'coreCF'];
-  function workbookCtrl($scope, config) {
+  workbookCtrl.$inject = ['$scope', 'workbookService', 'coreCF'];
+  function workbookCtrl($scope, workbookService, config) {
     var _spk = config.spreadKey;
     var that = this;
     that.workbook = null;
-console.info('-1');
+
+    that.toggle = toggle;
+
     // 变更监听
     $scope.$on(_spk.workbookChange, function(e, workbook) {
 console.warn('C工作簿更新!', workbook);
       that.workbook = workbook;
     });
+
+    // 切换表接口
+    function toggle(index) {
+      workbookService.toggle(index);
+    }
   }
 
 })();